requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
httpx==0.27.2
langchain==0.2.16
langchain-openai==0.1.22
python-dotenv==1.0.1
